% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{literal}
\alias{literal}
\title{Literal}
\usage{
literal(x)
}
\arguments{
\item{x}{Character, the string to be wrapped}
}
\value{
Character, the string wrapped, with beginliteral and endliteral tags and single quotes replacing
any internal double quotes.
}
\description{
To pass a template literal arrow function to Observable Plot, use the \code{literal} function to wrap the string.
For example to apply an axis tick format, you would use the \href{https://d3js.org/d3-format}{d3.format} library and
specify desired format using an arrow function, for example: \verb{d => d3.format("0.2f")(d)} applies two decimal places.
The option can be specified by wrapping the javascript arrow function in the R function as
\code{tickFormat=literal("d => d3.format('0.2f')(d)")}, note the use of single quotes
inside the double quotes.
Example app 4 illustrates some applications.
}
\examples{
literal("d => d3.format('0.2f')(d)")
}
