% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{obsjs_mark}
\alias{obsjs_mark}
\title{Observable Plot Mark}
\usage{
obsjs_mark(mark, data = NULL, options = NULL)
}
\arguments{
\item{mark}{Character the plot \href{https://observablehq.com/plot/features/marks}{mark} some common ones are:
\itemize{
\item dot: scatter plot dot, equivalent to geom_point
\item line: line plot, equivalent to geom_line
\item path: path plot, equivalent to geom_path
\item barX, barY: column or bar plots in the x or y dimension, equivalent to geom_col
\item text: scatter plot with text marks, equivalent to geom_text
\item ruleX, ruleY: horizontal or vertical lines across the plot panel
\item frame: outer box for the plot panel
\item tickX, tickY: strokes in the x or y dimension, of fixed length
}}

\item{data}{Data frame or vector of data to be plotted (R format)}

\item{options}{List of specification options for the mark including:
\itemize{
\item x axis variable \code{x},
\item y axis variable \code{y},
\item stroke color variable \code{stroke},
\item fill color variable \code{fill},
\item mark radius \code{r} for dot size,
\item faceting variable in the y dimension \code{fy}
\item faceting variable in the x dimension \code{fx}
\item text labels \code{text}
\item tip titles \code{title}
}}
}
\value{
Character, the mark data and specification
}
\description{
Takes a data frame, mark type and specification list and compiles a JSON format specification
mark specification for Observable Plot JS library.
}
\examples{
obsjs_mark("frame")  # add box to plot panel area
obsjs_mark("ruleY", data=0) # add x axis at y=0 
obsjs_mark("barY", data=data.frame(x=letters[1:5], y=1:5), options=list(x="x", y="y"))
}
