% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{obsjs_plot}
\alias{obsjs_plot}
\title{Observable Plot}
\usage{
obsjs_plot(inputId, ..., options = NULL)
}
\arguments{
\item{inputId}{Character, unique ID for plot in Shiny}

\item{...}{Plot marks as many as required, output from the \link{obsjs_mark}}

\item{options}{List of plot options}
}
\value{
A script module containing the plot attached to the DOM at the InputId.
}
\description{
Builds a \href{https://observablehq.com/plot/}{Observable Plot} from plot marks and a plot options list.
The output of this function can be called in Shiny through \link[shiny:renderUI]{shiny::renderUI} as HTML.
Shiny needs to import the javascript libraries for Observables, this can be done by including the
following two statements in the App UI:
\itemize{
\item \code{htmltools::HTML('<script src="https://cdn.jsdelivr.net/npm/d3@7"></script>')},
\item \code{htmltools::HTML('<script src="https://cdn.jsdelivr.net/npm/@observablehq/plot@0.6"></script>')},
}
}
\examples{
obsjs_plot("myplot",  obsjs_mark("barY", data=data.frame(x=letters[1:5], y=1:5), options=list(x="x", y="y")))
}
